+++ views/artists.hbs
@@ -1,4 +1,9 @@
 <div class="home-container">
+
+<div class='title'>
+    <h1>Search Results</h1>
+</div>
+
     <div class="row">
         {{#each artist}}
         <div class="album-card">
